requirements.txt
alembic==1.16.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.0
flask-cors==6.0.1
Flask-Migrate==4.1.0
flask-rebar==3.3.2
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyfingerprint==1.5
pyserial==3.5
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.38
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
